feature: create pool down methos to model Pool and user logic
