+++ app/Models/Pool.php
@@ -2,6 +2,7 @@
 
 namespace App\Models;
 
+use App\Exceptions\Pool\CompetitionMustBeUniqueInAPool;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
@@ -9,6 +10,7 @@ use Illuminate\Support\Collection;
 
 /**
  * @property Collection $poolInvitationsEmails
+ * @property string $name
  */
 class Pool extends Model
 {
@@ -46,5 +48,53 @@ class Pool extends Model
     public function poolInvitationsEmails()
     {
         return $this->hasMany(PoolInvitationsEmails::class);
+    }
+
+    public static function create(string $namePool)
+    {
+        $Pool = new static();
+
+        $Pool->name = $namePool;
+
+        if (! $Pool->save()) {
+            throw new \Exception('dont save Pool');
+        }
+
+        return $Pool;
+    }
+
+    public function addUser(User $User)
+    {
+        $this->users()->attach($User);
+    }
+
+    public function addCompetitions($competitions)
+    {
+        $this->haveCompetitionUnique($competitions);
+
+        $this->competitions()->attach($competitions);
+    }
+
+    protected function haveCompetitionUnique(iterable $competitions = null)
+    {
+        $competitionsCollect = collect($competitions);
+
+        $haveMoreThanOne = $competitionsCollect->count() > 1;
+
+        $haveUniqueCompetition = $competitionsCollect->filter(function (Competition $competition){
+                return $competition->must_be_unique;
+            })->count() > 0;
+
+        if($haveMoreThanOne && $haveUniqueCompetition)
+            throw new CompetitionMustBeUniqueInAPool('Have competition unique');
+    }
+
+    public function createInvitationsPoolEmails(iterable $emails)
+    {
+        foreach ($emails as $email){
+            $this->poolInvitationsEmails()->create([
+                'email' => $email,
+            ]);
+        }
     }
 }
